Stop throwing ApiError in the connector
These are not documented in the blob bundle and this break the responses in case
the blob bundle has its own error codes.

In case we want to communicate something to the blob bundle in the future we
need to throw our own exception type and make it part of the connector interface.