Fail if the target path doesn't exist and add a health check
In case the target is a mount to a different drive and the mount for some
reason failed we want to error out and don't just mkdir the path on the system,
potentially hiding any issues.

To avoid this check if the target path exists first and don't auto-create it.
In addition do this check in a health check, so we notice if the target
gets removed due to some error.

There is still a small race there where the file adding mkdirs the file right
after the check, but that would require larger changes.